requirements.txt
streamlit==1.37.1
pymongo==4.7.2
google-generativeai==0.7.2
python-dotenv==1.0.0
pandas>=2.0
